+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/configuration/ConfigurationEntries.kt
@@ -11,5 +11,5 @@ object ConfigurationEntries {
      */
     val IS_RESCAN_ENABLED = BooleanConfigurationEntry(ConfigKey("is_rescan_enabled"), true)
 
-    val IS_FLEXA_AVAILABLE = BooleanConfigurationEntry(ConfigKey("is_flexa_available"), false)
+    val IS_FLEXA_AVAILABLE = BooleanConfigurationEntry(ConfigKey("is_flexa_available"), true)
 }