+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this application adheres to [Semantic Versioning](https://semver.org/spec/v2
 
 ## [Unreleased]
 
+### Added
+- Flexa feature has been turned on
+- Disclaimer added to integrations screen
+
 ## [1.2.2 (789)] - 2024-11-18
 
 ### Added